// src/modules/catalog/controllers/dto/tag-query.dto.ts
import { createZodDto } from 'nestjs-zod';
import { z } from 'zod';
import { PaginationParamsSchema } from '@/common/schemas/pagination.schema';

const tagQuerySchema = PaginationParamsSchema.extend({
  search: z.string().optional(),
  id: z.uuid().optional(),
  name: z.string().optional(),
  isActive: z.enum(['true', 'false']).optional(),
  groupId: z.uuid().optional(),
  sortBy: z.enum(['createdAt', 'updatedAt']).optional().default('createdAt'),
  sortOrder: z.enum(['asc', 'desc']).optional().default('desc'),
});

export class TagQueryDto extends createZodDto(tagQuerySchema) {}